feat: add FreeRTOS and FreeRTOS-POSIX
This also includes a compatibility patch for the Pico SDK.
The pico_sync module defines a 'sem_init' method which clashes with
POSIX 'sem_init'.
This causes a 'multiple definitions' link issue when compiling with
FreeRTOS-POSIX and pico_sync.

As workaround the pico/sem.h methods are prefixed with 'pico_' and these
changes are included as local patch applied in the devcontainer.